--- src/applications/accredited-representative-portal/accreditation/21a/app-entry.jsx
@@ -1,14 +1,0 @@
-import 'platform/polyfills';
-import './sass/21a.scss';
-
-import startApp from 'platform/startup';
-
-import routes from './routes';
-import reducer from './reducers';
-import manifest from './manifest.json';
-
-startApp({
-  url: manifest.rootUrl,
-  reducer,
-  routes,
-});
