--- src/applications/accredited-representative-portal/accreditation/21a/components/YourCharacterReferencesDescription.jsx
@@ -1,24 +1,0 @@
-import React from 'react';
-
-const YourCharacterReferencesDescription = () => {
-  return (
-    <p>
-      Over the next couple of pages you’ll be asked to provide the full names,
-      addresses, email addresses, and current phone numbers of at least three
-      individuals who are not immediate family members and who have personal
-      knowledge of your character and qualifications to serve as an attorney.
-      <br />
-      <br />
-      You are encouraged to notify your character references that the Department
-      of Veterans Affairs Office of General Counsel will be sending an inquiry
-      for a character reference.
-      <br />
-      <br />
-      It is the applicant’s responsibility to ensure that responses are provided
-      to our office. If we do not receive at least three responses, your
-      application will not be processed.
-    </p>
-  );
-};
-
-export default YourCharacterReferencesDescription;
