+++ src/applications/accredited-representative-portal/accreditation/21a/containers/App.jsx
@@ -3,13 +3,19 @@ import PropTypes from 'prop-types';
 
 import RoutedSavableApp from 'platform/forms/save-in-progress/RoutedSavableApp';
 
+import Footer from '../../../components/common/Footer/Footer';
+import Header from '../../../components/common/Header/Header';
 import formConfig from '../config/form';
 
 const App = ({ location, children }) => {
   return (
-    <RoutedSavableApp formConfig={formConfig} currentLocation={location}>
-      {children}
-    </RoutedSavableApp>
+    <>
+      <Header />
+      <RoutedSavableApp formConfig={formConfig} currentLocation={location}>
+        {children}
+      </RoutedSavableApp>
+      <Footer />
+    </>
   );
 };
