--- src/applications/accredited-representative-portal/accreditation/21a/manifest.json
@@ -1,7 +1,0 @@
-{
-  "appName": "Apply to become a VA accredited attorney or claims agent",
-  "entryFile": "./app-entry.jsx",
-  "entryName": "21a",
-  "rootUrl": "/accreditation/attorney-claims-agent-form-21a",
-  "productId": "55b00994-6f3c-4527-98be-5a90e37eec01"
-}
