--- src/applications/accredited-representative-portal/accreditation/21a/routes.jsx
@@ -1,13 +1,0 @@
-import { createRoutesWithSaveInProgress } from 'platform/forms/save-in-progress/helpers';
-
-import formConfig from './config/form';
-import App from './containers/App';
-
-const route = {
-  path: '/',
-  component: App,
-  indexRoute: { onEnter: (_, replace) => replace('/introduction') },
-  childRoutes: createRoutesWithSaveInProgress(formConfig),
-};
-
-export default route;
