+++ src/applications/accredited-representative-portal/app-entry.jsx
@@ -3,11 +3,12 @@ import React from 'react';
 import startReactApp from '@department-of-veterans-affairs/platform-startup/react';
 
 import { Provider } from 'react-redux';
-import { BrowserRouter } from 'react-router-dom-v5-compat';
+import { Router, useRouterHistory } from 'react-router';
+import { createHistory } from 'history';
 
 import { connectFeatureToggle } from 'platform/utilities/feature-toggles';
 
-import routes from './routes';
+import route from './routes';
 import reducer from './reducers';
 import manifest from './manifest.json';
 import createReduxStore from './store';
@@ -20,8 +21,13 @@ connectFeatureToggle(store.dispatch);
 
 window.appName = manifest.entryName;
 
+// eslint-disable-next-line react-hooks/rules-of-hooks
+const history = useRouterHistory(createHistory)({
+  basename: manifest.rootUrl,
+});
+
 startReactApp(
   <Provider store={store}>
-    <BrowserRouter basename={manifest.rootUrl}>{routes}</BrowserRouter>
+    <Router history={history} routes={route} />
   </Provider>,
 );
