+++ src/applications/accredited-representative-portal/reducers/form21aSaveInProgress.js
@@ -1,6 +1,6 @@
 import { createSaveInProgressFormReducer } from 'platform/forms/save-in-progress/reducers';
 
-import formConfig from '../config/form';
+import formConfig from '../accreditation/21a/config/form';
 
 export default {
   form: createSaveInProgressFormReducer(formConfig),
